add: home app and modify urls
