+++ templates/shared/navbar2.html
@@ -4,7 +4,7 @@
   <div class="flex items-center justify-between px-4 mx-auto max-w-7xl h-14">
     <!-- Left Section: Logo -->
     <div class="flex items-center space-x-2">
-      <a href="{% url 'users:home' %}" class="flex items-center space-x-2">
+      <a href="{% url 'home:home' %}" class="flex items-center space-x-2">
         <img
           src="{% static 'images/logo.png' %}"
           alt="PicTrace Logo"
