+++ users/templates/users/password_reset_confirm.html
@@ -38,7 +38,7 @@
               name="new_password1"
               id="id_new_password1"
               class="block w-full px-3 py-2 mt-1 text-gray-700 placeholder-gray-400 bg-white border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
-              placeholder="請輸入新密碼"
+              placeholder="新密碼須包含英文及數字"
             />
           </div>
           <div>
