+++ users/urls.py
@@ -9,10 +9,8 @@ urlpatterns = [
     path("login/", views.login_view, name="login"),
     path("register/", views.register_view, name="register"),
     path("logout/", views.logout_view, name="logout"),
-    path("", views.home_view, name="home"),
     path("upload_avatar/", views.upload_avatar, name="upload_avatar"),
     path("profile/", views.profile_view, name="profile"),
-    path("activate/<uidb64>/<token>/", views.activate_account, name="activate"),
     path(
         "password_reset/",
         views.CustomPasswordResetView.as_view(),
